==> src/step1_process_listings.ipynb <==
# Notebook 1: Process Listings Dataset 
dice_com-job_us_sample.csv  
https://www.kaggle.com/datasets/PromptCloudHQ/us-technology-jobs-on-dicecom
#### This notebook produces the following data into the _output_datasets_ folder:
```
(LISTING) NODE						listing__node.csv
listing_id:ID
listing_title
description
:LABEL = "LISTING"

[NEEDS] RELATION					needs__relation.csv
:START_ID = listing_id
:END_ID = skill_id
:TYPE = "NEEDS"

[LOCATED_IN] RELATION					located_in__relation.csv
:START_ID = listing_id
:END_ID = location_id
:TYPE = "LOCATED_IN"

(LOCATION) NODE						location__node.csv
location_id:ID
location_name
:LABEL = "LOCATION"

[POSTED] RELATION					posted__relation.csv
:START_ID = company_id
:END_ID = listing_id
:TYPE = "POSTED"

(COMPANY) NODE						company__node.csv
company_id:ID
company_name
:LABEL = "COMPANY"
```
#### Also, it produces intermediate datasets, used for further Skill Matching steps into the _temp_datasets_ folder:
```
(LISTING_SKILL) NODE					listing_skills_TEMP.csv
listing_skill_id
listing_skill_name
```



# Setup

# %pip install stanza
# %pip install spacy
# %pip install nltk
# !python -m spacy download en_core_web_sm

from pathlib import Path
import pandas as pd
import numpy as np
import stanza
import spacy
import re

stanza.download('en') 
nlp_spacy = spacy.load("en_core_web_sm")
nlp_stanza = stanza.Pipeline('en', processors='tokenize, ner', use_gpu=False, pos_batch_size=3000, download_method=None)

# this cell is to support running the notebook in Google Colab

mydrive = ""  # this is when we run locally

# Google Colab:
# from google.colab import drive
# drive.mount('/content/drive')
# mydrive = "/content/drive/MyDrive/DSE 203 — etl/DSE203_Project/"  # this is when we run on COLAB Leslie
# mydrive = "/content/drive/MyDrive/DSE203_Project/"  # this is when we run on COLAB Sergey



Path("output_datasets").mkdir(parents=True, exist_ok=True)
Path("temp_datasets").mkdir(parents=True, exist_ok=True)


input_dir = mydrive+"input_datasets/"
output_dir = mydrive+"output_datasets/"
temp_dir = mydrive+"temp_datasets/"

### Import data

# SMALL, SAMPLED DATASET FOR TESTING:
# df = pd.read_csv(input_dir+"dice_small.csv")

# FULL DATASET:
df = pd.read_csv(input_dir+"dice_com-job_us_sample.csv")

df.info()
df.head(3)

### Set up NER to extract skills

def extract_entities_stanza(series):
    '''
    apply stanza to extract ORG and PRODUCT entities
    '''
    
    doc = nlp_stanza(series)
    entities_skills = doc.entities
    
    result = list({x.text for x in entities_skills if (x.type == 'ORG') or (x.type == 'PRODUCT')})
    
    return result

def extract_entities_spacy(series):
    '''
    apply spacy to extract ORG and PRODUCT entities
    '''

    doc = nlp_spacy(series, disable=["tok2vec", "parser"])
    entities_skills = doc.ents
    
    result = list({x.text for x in entities_skills if (x.label_ == 'ORG') or (x.label_ == 'PRODUCT')})
    
    return result

# Prep

### Only save columns we want

def basic_cleanup(df: pd.DataFrame):
    '''
    drop cols that are not useful to us
    '''
    to_drop = ['advertiserurl', 'employmenttype_jobstatus', 'jobid', 'uniq_id', 'postdate', 'shift', 'site_name']
    df = df.drop(to_drop, axis=1) \
           .drop_duplicates(subset=['company', 'joblocation_address', 'jobtitle']) \
           .rename(columns={'company':'company_name','jobdescription':'description','joblocation_address':'location_name','jobtitle':'listing_title'})
    return df

df = basic_cleanup(df)

df.info()
df.head(3)

### Clean `skills` as strings

def clean_skills(string):
    '''
    remove everything but letters, numbers, commas, +, #, and regular punctuation
    '''

    if type(string)!=str:
      return np.nan
    string = string.replace(' / ', ' ') \
                   .replace('/', ' ') \
                   .replace('... ', ', ') \
                   .replace('...', ', ')
    string = string.replace('  ', ', ')
    string = re.sub('[^a-zA-Z0-9,?!+# ]+', '', string)
    return string

def process_non_skills(skillcol):
    '''
    use empty strings (to fill later with jobdescription skills)
    '''

    mask = (skillcol.str.lower()=="null") | \
           (skillcol.str.lower().str.contains("see below")) | \
           (skillcol.str.lower()=="please see job description") | \
           (skillcol.str.lower()=="see job description") | \
           (skillcol.str.lower()=="see job overview") | \
           (skillcol.str.lower()=="full time") | \
           (skillcol.str.lower()=="please refer to job description") | \
           (skillcol.str.startswith("TAD PGS, INC specializes in")) #known offender of 60+ rows
    skillcol[mask] = ''
    return skillcol

df['skills'] = df['skills'].apply(clean_skills)
df['skills'] = process_non_skills(df['skills'])

df.info()
df.head(3)

### Clean `Job Descriptions` only for rows that doesn't have any valid skills

def clean_job_description(string):
    '''
    remove everything but letters, numbers, commas, +, #, and regular punctuation
    '''
    
    if type(string)!=str:
      return np.nan
    string = string.replace('•','. ') \
                   .replace('\n','. ') \
                   .replace('...', '. ') \
                   .replace('\xa0','. ') \
                   .replace('\t', ' ')
    string = string.replace('  ', ' ')
    string = re.sub('[^a-zA-Z0-9,.?!+# ]+', '', string)
    return string

'''clean only job descriptions that don't have valid skills'''
df.description = df.apply(lambda row: clean_job_description(row.description) if row['skills']=='' else row.description, axis=1)

df.info()
df.head(3)

# Format the `skills` column how we want it

### Convert `skills` string to list

def skills_to_list(df: pd.DataFrame):
    '''
    split lists along commas
    '''
    
    df = df.dropna()
    df['skills'] = df['skills'].apply(lambda x: x.split(',') if x!='' else [])
    return df

df = skills_to_list(df)

df.info()
df.head(3)

### Clean `skills` as lists

Use NER to extract skills from longer sentences that were incorrectly save within the `skills` lists

def clean_list(row, threshold: int=6):
    '''
    strip whitespaces and get rid of empty entries
    also get rid of "skills" that are really just sentences, but extract ORG and PRODUCT from them first
    '''
    
    row = [r.strip() for r in row if r]
    
    '''extract ORG and PRODUCT from sentences'''
    one = [extract_entities_stanza(r) for r in row if len(r.split()) >= threshold] #using stanza
    two = [extract_entities_spacy(r) for r in row if len(r.split()) >= threshold] #using spacy
    '''combine with existing list of skills (without duplicates)'''
    x = one + two
    row = list(set(sum(x, row)))
    
    row = [r for r in row if (len(r.split()) < threshold)] #get rid of sentences in the list
    return row

def post_ner_fix(row):
    '''
    get rid of "travel" and "us government" skills that are leftover from improperly selected skills from using NER
    (things like "50 tavel" and "US Government Secret")
    '''
    row = [x for x in row if ("travel" not in x.lower() and "us government" not in x.lower())]
    return row

%%time
df['skills'] = df['skills'].apply(clean_list)
df['skills'] = df['skills'].apply(post_ner_fix)

df.info()
df.head(3)

### Apply NER to extract skills from Job Description column, but only if we are missing them in Skills column

def extend_lists(df):
    '''
    This function will get the main dataframe and will combine lists of skills 
    from different columns, will remove duplicates and then will produce a final skills list.
    '''
    
    one  = df['description_stanza']
    two = df['description_spacy']
    three = df['skills']
    
    result = one + two + three

    # lowercase all skills
    result = [x.lower() for x in result]
    result = list(set(result))
    
    return result

%%time

'''extract with stanza'''
df['description_stanza'] = np.empty((len(df), 0)).tolist()
df['description_stanza'] = df.apply(lambda row: extract_entities_stanza(row.description) if row['skills']==[] else row.description_stanza, axis=1)

'''extract with spacy'''
df['description_spacy'] = np.empty((len(df), 0)).tolist()
df['description_spacy'] = df.apply(lambda row: extract_entities_spacy(row.description) if row['skills']==[] else row.description_stanza, axis=1)

'''combine everything together and remove duplicate skills'''
df['listing_skill_name'] = df.apply(extend_lists, axis=1)

df

############### Sergey: this removes some good stuff too, like [unix] ##################################

def remove_empty_skills(df: pd.DataFrame):
    '''
    get rid of all skill-lists that are 1 item long
    '''
    
    df['skills_count'] = df['listing_skill_name'].apply(lambda x: len(x))
    df = df[df['skills_count']>1] \
          .drop(columns=['skills_count'])
    return df

'''last bit of pruning empty skillsets (and skillsets of size 1)'''
df = remove_empty_skills(df)

'''drop rows with empty listing_skill_name (empty list or null)'''
df = df[df['listing_skill_name'].astype(bool)] \
       .dropna(subset=['listing_skill_name']) \
       .reset_index(drop=True).reset_index() \
       .rename(columns={'index':'listing_id:ID'})

df

'''save a copy just in case'''
df[['listing_id:ID', 'company_name', 'description', 'location_name', 'listing_title', 'listing_skill_name']] \
  .to_csv(temp_dir+"templistingdataframe.csv", index=False)

# Saving nodes and relations as CSVs for Neo4j

### (node) LISTING

listing_df = df[['listing_id:ID','listing_title','description']].copy()
listing_df[':LABEL'] = "LISTING"
listing_df

listing_df.to_csv(output_dir+"listing__node.csv", index=False)

### (temporary node) LISTING_SKILL

listing_skill_df = df[['listing_skill_name']].copy() \
                     .explode('listing_skill_name') \
                     .drop_duplicates() \
                     .reset_index(drop=True).reset_index() \
                     .rename(columns={'index':'listing_skill_id'})

listing_skill_df

listing_skill_df.to_csv(temp_dir+"listing_skills_TEMP.csv", index=False)

### (node) LOCATION

location_df = df[['location_name']].copy() \
                .drop_duplicates() \
                .reset_index(drop=True).reset_index() \
                .rename(columns={'index':'location_id:ID'})
location_df[':LABEL'] = "LOCATION"
location_df

location_df.to_csv(output_dir+"location__node.csv", index=False)

### (node) COMPANY

company_df = df[['company_name']].copy() \
                .drop_duplicates() \
                .reset_index(drop=True).reset_index() \
                .rename(columns={'index':'company_id:ID'})
company_df[':LABEL'] = "COMPANY"
company_df

company_df.to_csv(output_dir+"company__node.csv", index=False)

### [relation] NEEDS

needs_df = df[['listing_id:ID', 'listing_skill_name']].copy()
needs_df = needs_df.explode('listing_skill_name') \
                   .dropna() \
                   .merge(listing_skill_df, on='listing_skill_name') \
                   .drop(columns=['listing_skill_name']) \
                   .rename(columns={'listing_id:ID':':START_ID', 'listing_skill_id':':END_ID'})
needs_df[':TYPE'] = "NEEDS"
needs_df

needs_df.to_csv(output_dir+"needs__relation.csv", index=False)

### [relation] LOCATED_IN

located_in_df = df[['listing_id:ID', 'location_name']].copy()
located_in_df = located_in_df.dropna() \
                   .merge(location_df, on='location_name') \
                   .drop(columns=['location_name', ':LABEL']) \
                   .rename(columns={'listing_id:ID':':START_ID', 'location_id:ID':':END_ID'})
located_in_df[':TYPE'] = "LOCATED_IN"
located_in_df

located_in_df.to_csv(output_dir+"located_in__relation.csv", index=False)

### [relation] POSTED

posted_df = df[['listing_id:ID', 'company_name']].copy()
posted_df = posted_df.dropna() \
                   .merge(company_df, on='company_name') \
                   .drop(columns=['company_name', ':LABEL']) \
                   .rename(columns={'company_id:ID':':START_ID', 'listing_id:ID':':END_ID'})
posted_df[':TYPE'] = "POSTED"
posted_df

posted_df.to_csv(output_dir+"posted__relation.csv", index=False)
