==> src/step2_process_courses.ipynb <==
# Notebook 2: Process Courses Dataset 
Coursera.csv  
https://www.kaggle.com/datasets/khusheekapoor/coursera-courses-dataset-2021

#### This notebook produces the following data into the _output_datasets_ folder:
```
(COURSE) NODE						course__node.csv
course_id:ID
course_name
course_difficulty_level
course_url
:LABEL = "COURSE”
```
#### Also, it produces intermediate datasets, used for further Skill Matching steps into the _temp_datasets_ folder:
```
(COURSE_SKILL) NODE					courses_skills_TEMP.csv
course_skill_id
course_skill_name

(COURSE_SKILL) RELATION					courses_skills_relationship_TEMP.csv
course_id
course_skill_id
```



## Imports

# %pip install stanza
# %pip install spacy
# %pip install nltk
# !python -m spacy download en_core_web_sm

import pandas as pd
import numpy as np
import stanza
import spacy
import re
stanza.download('en') 
nlp_spacy = spacy.load("en_core_web_sm")
nlp_stanza = stanza.Pipeline('en', processors='tokenize, ner', use_gpu=False, pos_batch_size=3000, download_method=None)

# this cell is to support running the notebook in Google Colab

mydrive = ""  # this is when we run locally

# Google Colab:
# from google.colab import drive
# drive.mount('/content/drive')
# mydrive = "/content/drive/MyDrive/DSE 203 — etl/DSE203_Project/"  # this is when we run on COLAB Leslie
# mydrive = "/content/drive/MyDrive/DSE203_Project/"  # this is when we run on COLAB Sergey

input_dir = mydrive+"input_datasets/"
output_dir = mydrive+"output_datasets/"
temp_dir = mydrive+"temp_datasets/"

## Read Data

# SMALL, SAMPLED DATASET FOR TESTING:
# courses_df = pd.read_csv(input_dir+'coursera_small.csv')

# FULL DATASET:
courses_df = pd.read_csv(input_dir+'Coursera.csv')

courses_df.info()

courses_df.rename(columns = {"Course Name":"course_name",
                     "Course Rating": "course_rating",
                     "Difficulty Level": "course_difficulty_level",
                     "Course URL": "course_url",
                     "Course Description": "course_description",
                     "Skills": "course_skills"
                    }, inplace=True)

columns_to_leave = ["course_name", 
                    "course_description",
                    "course_difficulty_level",
                    "course_url",
                    "course_skills"]
courses_df = courses_df[columns_to_leave]
courses_df.columns

## Text Cleaning

def clean_text(string):
    '''
    remove everything but the alphabetic letters
    '''
    string = re.sub('[^a-zA-Z,.?! ]+', '', string)
    return string


columns_to_clean = ['course_name', 'course_description', 'course_skills']

for column in columns_to_clean:
    courses_df[column] = courses_df[column].apply(clean_text)

### Convert Skills string to list

courses_df.course_skills = courses_df.course_skills.str.split('  ')
courses_df.head(3)

## Apply NER to extract skills

def extract_entities_stanza(series):
    '''
    This function will get a dataframe column (series) and will extract skills from the text
    using Stanza library.
    '''
    
    doc = nlp_stanza(series)
    entities_skills = doc.entities
    
    result = list({x.text for x in entities_skills if (x.type == 'ORG') or (x.type == 'PRODUCT')})
    
    return result

def extract_entities_spacy(series):
    '''
    This function will get a dataframe column (series) and will extract skills from the text
    using Spacy library.
    '''

    doc = nlp_spacy(series, disable=["tok2vec", "parser"])
    entities_skills = doc.ents
    
    result = list({x.text for x in entities_skills if (x.label_ == 'ORG') or (x.label_ == 'PRODUCT')})
    
    return result

def extend_lists(df):
    '''
    This function will get a Courses dataframe and will combine lists of skills 
    from different columns, will remove duplicates and then will produce a final skills list.
    '''
    # for simplicity
    one  = df.course_name_stanza
    two = df.course_name_spacy
    three = df.description_stanza
    four = df.description_spacy
    five = df.course_skills
    
    # combine into one and eliminate dups
    result = one + two + three + four + five
    
    # lowercase all skills
    result = [x.lower() for x in result]
    
    result = list(set(result))
    
    return result

%%time

# extract with Stanza
courses_df['course_name_stanza'] = courses_df['course_name'].apply(extract_entities_stanza)
courses_df['description_stanza'] = courses_df['course_description'].apply(extract_entities_stanza)

# extract with Space
courses_df['course_name_spacy'] = courses_df['course_name'].apply(extract_entities_spacy)
courses_df['description_spacy'] = courses_df['course_description'].apply(extract_entities_spacy)

# Combine everything together and remove duplicate skills
courses_df['all_course_skills'] = courses_df.apply(extend_lists,axis=1)

courses_df.head(3)

## Create relational tables

### Explode table by all skills

courses_exploded_df = courses_df[['course_name', 'all_course_skills']].explode('all_course_skills')

courses_exploded_df.columns = ['course_name', 'course_skill']
courses_exploded_df = courses_exploded_df.drop_duplicates(subset=['course_name','course_skill'])
courses_exploded_df.reset_index(inplace=True)
courses_exploded_df.columns = ['course_id', 'course_name', 'course_skill']
courses_exploded_df[courses_exploded_df.course_skill.str.contains('ython')]

## Create table of course skills (without duplicates)

courses_skills_df = courses_exploded_df[['course_skill']].copy()
courses_skills_df = courses_skills_df.drop_duplicates() \
               .reset_index(drop=True) \
               .reset_index() \
               .rename(columns={'course_skill':'course_skill_name', 'index':'course_skill_id'})
courses_skills_df.course_skill_id = courses_skills_df.course_skill_id.astype('int')

courses_skills_df[courses_skills_df.course_skill_name.str.contains('ython')]

courses_skills_df.to_csv(temp_dir+"courses_skills_TEMP.csv", index=False)

## Map course skills back to the courses to for creating the intermedite (temp) relationships table

courses_exploded_df = courses_exploded_df.merge(courses_skills_df, how='outer', left_on='course_skill', right_on='course_skill_name')
courses_exploded_df

courses_exploded_df[["course_id", "course_skill_id"]].to_csv(temp_dir+"courses_skills_relationship_TEMP.csv",index=False)

### (node) COURSE

columns_to_leave = ["course_name", 
                    "course_difficulty_level",
                    "course_url"]
courses_df = courses_df[columns_to_leave]
courses_df[":LABEL"] = "COURSE"
courses_df

courses_df.to_csv(output_dir+"course__node.csv", index_label="course_id:ID")
