==> src/step4_create_occupation_related_nodes.ipynb <==
# Notebook 4: Download Occupations and Carrer Outlook Data from ONET API and Produce Nodes and Relations:

```
(OCCUPATION) NODE						occupation__node.csv
occupation_id:ID
occupation_title
onet_code
occupation_synonyms[]
occupation_description
occupation_salary
:LABEL = "OCCUPATION"

[HAS_FUTURE] RELATION					has_future__relation.csv
:START_ID = occupation_id
:END_ID = career_outlook_id
:TYPE = "HAS_FUTURE"

(CAREER_OUTLOOK) NODE					career_outlook__node.csv
career_outlook_id:ID
career_outlook
:LABEL = "CAREER_OUTLOOK"
```

## Imports

# %pip install xmltodict

import requests
import json
import xmltodict
import pandas as pd

# this cell is to support running the notebook in Google Colab

mydrive = ""  # this is when we run locally

# Google Colab:
# from google.colab import drive
# drive.mount('/content/drive')
# mydrive = "/content/drive/MyDrive/DSE 203 — etl/DSE203_Project/"  # this is when we run on COLAB Leslie
# mydrive = "/content/drive/MyDrive/DSE203/DSE203_Project/"  # this is when we run on COLAB Jessica
# mydrive = "/content/drive/MyDrive/DSE203_Project/"  # this is when we run on COLAB Sergey

input_dir = mydrive+"input_datasets/"
output_dir = mydrive+"output_datasets/"
temp_dir = mydrive+"temp_datasets/"

# !!! NOTE !!!
## The data has been already downloaded and can be read from 'input_data' folder. You can directly go to cell

[SKIP DOWNLOAD](#read)

## Download Data from ONET API

#Request from ONET API all occupations

onet_username = '<ONET_USERNAME>' 
onet_password = '<ONET_PASSWORD>'

response = requests.get('https://services.onetcenter.org/ws/online/occupations?start=1&end=1020', auth=(onet_username, onet_password))
job_dict = xmltodict.parse(response.content)

job_codes = []
job_titles = []
for code in job_dict['occupations']['occupation']:
    job_codes += [code['code']]
    job_titles += [code['title']]
      
#dataframe of jobs
occupation_df = pd.DataFrame({'job_code': job_codes, 
                            'occupation_title': job_titles}) 


occupation_df.tail(5)

# For each occupation, grab the career descriptions, synonyms, expected salary and the career outlook

occupation_synonyms = []
occupation_description = []
occupation_growth = []
occupation_salary = []

occ_len = len(occupation_df['job_code'])

for i, job_code in enumerate(occupation_df['job_code']):
    
    career_response = requests.get(f'https://services.onetcenter.org/ws/online/occupations/{job_code}/summary', auth=(onet_username, onet_password))
    print(f"\rGot response for Summary {i} out of {occ_len}", end="")
    career_dict = xmltodict.parse(career_response.content)
    
    if 'summary_report' in career_dict and 'occupation' in career_dict['summary_report'] and 'sample_of_reported_job_titles' in career_dict['summary_report']['occupation']:
        occupation_synonyms += [career_dict['summary_report']['occupation']['sample_of_reported_job_titles']['title']]
    else:
        occupation_synonyms += ['']

    if 'summary_report' in career_dict:
        occupation_description += [career_dict['summary_report']['occupation']['description']]
    else:
        occupation_description += ['']
        
    report_response = requests.get(f'https://services.onetcenter.org/ws/mnm/careers/{job_code}/report', auth=(onet_username, onet_password))
    print(f"\rGot response for Reports {i} out of {occ_len}", end="")
    report_dict = xmltodict.parse(report_response.content)
    if 'report' in report_dict:
        occupation_growth += [report_dict['report']['job_outlook']['outlook']['category']]
        if 'salary' in report_dict['report']['job_outlook']:
            if 'annual_median' in report_dict['report']['job_outlook']['salary']:
                occupation_salary += [report_dict['report']['job_outlook']['salary']['annual_median']]
            else:
                occupation_salary += ['']
        else:
            occupation_salary += ['']
    else: 
        occupation_growth += ['']
        occupation_salary += ['']


#Put all responses into a dataframe
occupation_df['occupation_synonyms'] = occupation_synonyms
occupation_df['occupation_description'] = occupation_description
occupation_df['occupation_growth'] = occupation_growth
occupation_df['occupation_salary'] = occupation_salary
occupation_df.reset_index(inplace=True)
occupation_df = occupation_df.rename(columns={"index": "occupation_id:ID"})
occupation_df.head(5)

# save to csv, so we don't need to re-download
# occupation_df.to_csv(input_dir+'occupations_dataset_downloaded.csv',index=False)

# RUN THIS CELL IF YOU WANT TO SKIP THE DOWNLOAD PART <a id="read"></a>

occupation_df = pd.read_csv(input_dir+'occupations_dataset_downloaded.csv')
occupation_df.head(3)

## Save to Files

## Create job outlook nodes by finding unique values and assigning an id
career_outlook = occupation_df.occupation_growth.unique()
career_outlook_df = pd.DataFrame({'career_outlook': career_outlook}).dropna()
career_outlook_df = career_outlook_df[career_outlook_df.career_outlook != '']
career_outlook_df.reset_index(inplace=True)
career_outlook_df = career_outlook_df.rename(columns={"index": "career_outlook_id:ID", "career_growth": "career_outlook"})
career_outlook_df[':LABEL'] = "CAREER_OUTLOOK"

career_outlook_df.to_csv(output_dir+'career_outlook__node.csv', index=False)
career_outlook_df.tail(4)

## create has_future relation by finding unique values and assigning an id
career_outlook_df = career_outlook_df.rename(columns={ "career_outlook": "occupation_growth"})
has_future_df = pd.merge(occupation_df, career_outlook_df, on='occupation_growth', how='inner')
has_future_df[':TYPE'] = 'HAS_FUTURE'
has_future_df = has_future_df[['occupation_id:ID', 'career_outlook_id:ID', ':TYPE']]
has_future_df = has_future_df.rename(columns={ 'occupation_id:ID': ":START_ID", 'career_outlook_id:ID': ":END_ID"})
# has_future_df.drop(has_future_df.columns[1], axis=1, inplace=True)
has_future_df.head(5)

has_future_df.to_csv(output_dir+'has_future__relation.csv',  index=False)
has_future_df.head(2)

occupation_df = occupation_df.loc[:, occupation_df.columns != 'occupation_growth']
occupation_df = occupation_df.rename(columns={"job_code": "onet_code"})
occupation_df[':LABEL'] = 'OCCUPATION'
occupation_df.tail(5)

occupation_df.to_csv(output_dir+'occupation__node.csv', index=False)
